Visualisation viewsheds -- Work In progress
