+++ essai/dem_file_projected.tif.aux.xml
@@ -1,0 +1,12 @@
+<PAMDataset>
+  <PAMRasterBand band="1">
+    <Metadata>
+      <MDI key="STATISTICS_APPROXIMATE">YES</MDI>
+      <MDI key="STATISTICS_MINIMUM">0</MDI>
+      <MDI key="STATISTICS_MAXIMUM">1317</MDI>
+      <MDI key="STATISTICS_MEAN">273.81396153415</MDI>
+      <MDI key="STATISTICS_STDDEV">215.8385975778</MDI>
+      <MDI key="STATISTICS_VALID_PERCENT">100</MDI>
+    </Metadata>
+  </PAMRasterBand>
+</PAMDataset>
